Determine expected best card at depth 4 using Monte Carlo rollouts

+++ C++/include/search.h
@@ -7,11 +7,11 @@
 enum TTType {UPPER, LOWER, EXACT};
 
 struct TTEntry {
-    uint8_t score;
+    int score;
     TTType type;
     uint8_t best_move;
     uint8_t trick_depth;
     uint64_t hash;
 };
 
-uint8_t minimax(GameState& game_state, uint8_t trick_depth, uint8_t alpha, uint8_t beta, bool maximizing, uint8_t ply);
+int minimax(GameState& game_state, uint8_t trick_depth, int alpha, int beta, bool maximizing, uint8_t ply);
